src/requirements.txt
functions-framework==3.8.1
python-telegram-bot==21.10
requests==2.32.3
pydantic==2.10.5